+++ app/[locale]/rest-client/[[...params]]/page.tsx
@@ -1,5 +1,9 @@
 import { Request } from '@components';
 
 export default function RestClient() {
-  return <Request />;
+  return (
+    <div>
+      <Request />
+    </div>
+  );
 }
