--- src/components/Body/index.ts
@@ -1,1 +1,0 @@
-export { Body } from './Body';
