--- src/components/GeneratedCode/index.ts
@@ -1,1 +1,0 @@
-export { GeneratedCode } from './GeneratedCode';
