--- src/components/Headers/index.ts
@@ -1,1 +1,0 @@
-export { Headers } from './Headers';
