+++ src/components/index.ts
@@ -1,2 +1,1 @@
 export { Request } from './Request';
-export { Button } from './ui';
